+++ net/scm/Kconfig
@@ -3,9 +3,6 @@
 #
 
 config SCM
-	tristate "SCM Network Device Support"
-	---help---
-	  To compile this driver as a module, choose M here: the module
-	  will be called scm. If unsure, say N.
+	bool "SCM Network Device Support"
