+++ net/scm/Makefile
@@ -3,6 +3,5 @@
 # Makefile for the Linux X.25 Packet layer.
 #
 
-obj-$(CONFIG_SCM) += scm.o
-
-scm-y			:= scm.o
+obj-$(CONFIG_SCM) += xscm.o
+xscm-y := xscm.o scm_proxy.o
