+++ net/scm/xscm.c
